TestNinja/Mocking/VideoServiceDIConstructorInjection.cs
using Newtonsoft.Json;
using System;
using System.Collections.Generic;

namespace TestNinja.Mocking
{
    public class VideoServiceDIConstructorInjection
    {
        private IFileReader _fileReader;
        private IVideoRepository _videoRepository;

        public VideoServiceDIConstructorInjection(IFileReader fileReader = null, IVideoRepository videoRepository = null)
        {
            _fileReader = fileReader ?? new FileReader();
            _videoRepository = videoRepository ?? new VideoRepository();
        }

        public string ReadVideoTitle()
        {
            var str = _fileReader.Read("video.txt");
            var video = JsonConvert.DeserializeObject<Video>(str);
            if (video == null)
                return "Error parsing the video.";
            return video.Title;
        }

        public string GetUnprocessedVideosAsCsv()
        {
            var videoIds = new List<int>();
            var videos = _videoRepository.GetUnprocessedVideos();

            foreach (var v in videos)
                videoIds.Add(v.Id);

            return String.Join(",", videoIds);
        }
    }
}